api/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
redis
qdrant-client
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
bcrypt==4.0.1